+++ src/app/vulnerability-db/page.tsx
@@ -1,13 +1,7 @@
-"use client";
-
-import Link from "next/link";
-
 import {
   Card,
-  CardContent,
   CardCoverImage,
   CardFooter,
-  CardHeader,
   CardLinkButton,
   CardProps,
   CardSubTitle,
@@ -15,136 +9,16 @@ import {
   CardTitleProps,
   CardTitleWrapper,
 } from "@/components/ui/Card";
-import { IconExternalLink, IconPin } from "@/components/ui/Icons";
-import { Label } from "@/components/ui/Label";
-import Pagination from "@/components/ui/Pagination";
-import { Ranking } from "@/components/ui/Ranking";
+import Dashboard from "@/components/vulnerability-db/Dashboard";
+import { vulnerabilityDBDatas } from "@/lib/dummyVulns";
 import { cn } from "@/lib/utils";
-import { useState } from "react";
-
-const vulnerabilityDBDatas = [
-  {
-    id: 1,
-    title: "[취약성 경고] Microsoft의 여러 보안 취약점에 대한 CNNVD의 보고서",
-    content:
-      "최근 Microsoft에서 발표한 보안 공지에 따르면, 원격 코드 실행, 권한 상승 등 다양한 취약점이 포함되어 있으며, 특히 Windows 10 운영체제에서 위험성이 큰 것으로 평가되었습니다.",
-    createdAt: "2024.03.08 13:30:24",
-  },
-  {
-    id: 2,
-    title: "2023년 12월 CNNVD 호환 서비스 신제품 발표",
-    content:
-      "CNNVD는 최신 위협 정보를 신속하게 분석하고 대응할 수 있는 새로운 서비스 플랫폼을 발표했습니다. 이 플랫폼은 기업 보안 관리자를 위한 실시간 경고 시스템을 제공합니다.",
-    createdAt: "2023.12.01 09:00:00",
-  },
-  {
-    id: 3,
-    title:
-      "[취약성 보고서] CISCO IOS XE 소프트웨어의 보안 취약점에 대한 CNNVD의 보고서",
-    content:
-      "CISCO의 IOS XE 소프트웨어에서 발견된 취약점은 원격 공격자가 관리자 권한을 획득할 수 있게 해주며, 네트워크 인프라 전반에 걸쳐 심각한 위협을 초래할 수 있습니다.",
-    createdAt: "2024.01.15 16:45:30",
-  },
-  {
-    id: 4,
-    title:
-      "[취약성 경고] Apache Dubbo의 다양한 보안 취약점에 대한 CNNVD의 보고서",
-    content:
-      "Apache Dubbo는 인기 있는 분산 시스템 프레임워크로, 최근 원격 코드 실행과 서비스 거부 공격에 취약한 것으로 밝혀졌습니다. 최신 패치를 적용하는 것이 강력히 권장됩니다.",
-    createdAt: "2024.02.20 11:22:10",
-  },
-  {
-    id: 5,
-    title: "[취약성 경고] Adobe Acrobat Reader의 치명적 취약점",
-    content:
-      "Adobe Acrobat Reader에서 발견된 이 취약점은 악성 PDF 파일을 통해 원격 코드 실행을 허용할 수 있으며, 특히 개인 사용자와 기업 모두에게 큰 위험을 초래합니다.",
-    createdAt: "2024.03.01 14:15:00",
-  },
-  {
-    id: 6,
-    title: "[취약성 경고] Google Chrome의 제로데이 취약점",
-    content:
-      "Google Chrome에서 발견된 제로데이 취약점은 공격자가 브라우저 세션을 장악할 수 있는 가능성을 제공합니다. 사용자들은 즉시 브라우저를 최신 버전으로 업데이트할 필요가 있습니다.",
-    createdAt: "2024.03.10 10:30:45",
-  },
-  {
-    id: 7,
-    title: "CNNVD 연례 보고서 발표: 2023년의 주요 보안 위협 분석",
-    content:
-      "2023년 한 해 동안 발생한 주요 보안 위협을 분석한 CNNVD 연례 보고서가 발표되었습니다. 이 보고서에는 랜섬웨어 공격의 증가, IoT 기기의 취약성 등이 주요 내용으로 다뤄졌습니다.",
-    createdAt: "2024.01.05 08:00:00",
-  },
-  {
-    id: 8,
-    title: "[취약성 보고서] Oracle WebLogic Server의 취약점",
-    content:
-      "Oracle WebLogic Server에서 발견된 취약점은 공격자가 악성 코드를 주입할 수 있는 가능성을 제공하며, 이로 인해 기업 내 서버가 위험에 처할 수 있습니다.",
-    createdAt: "2024.02.18 17:05:22",
-  },
-  {
-    id: 9,
-    title: "[취약성 경고] VMware vSphere의 보안 문제",
-    content:
-      "VMware vSphere 환경에서 발생하는 이 보안 문제는 가상화 서버를 표적으로 하는 공격이 가능하게 만들며, 특히 클라우드 인프라를 사용하는 조직들에게 큰 위협이 됩니다.",
-    createdAt: "2024.02.28 12:40:00",
-  },
-  {
-    id: 10,
-    title: "[취약성 경고] Microsoft Exchange Server의 취약점",
-    content:
-      "Microsoft Exchange Server에서 발견된 이 취약점은 원격에서 서버를 제어할 수 있게 만들며, 특히 이메일 시스템을 표적으로 하는 공격에 취약합니다.",
-    createdAt: "2024.03.08 13:30:24",
-  },
-  {
-    id: 11,
-    title: "[취약성 경고] Apple iOS의 보안 결함",
-    content:
-      "Apple의 iOS 운영체제에서 발견된 이 보안 결함은 사용자 데이터의 유출을 초래할 수 있으며, 악성 앱이 기기에 무단으로 접근할 수 있게 만듭니다.",
-    createdAt: "2024.03.02 15:50:10",
-  },
-  {
-    id: 12,
-    title: "[취약성 경고] Linux Kernel의 심각한 취약점",
-    content:
-      "Linux Kernel에서 발견된 이 심각한 취약점은 루트 권한을 가진 공격자가 시스템 전체를 장악할 수 있게 해줍니다. 시스템 관리자는 즉각적인 패치가 필요합니다.",
-    createdAt: "2024.03.11 09:20:55",
-  },
-];
+import Link from "next/link";
 
 export default function VulnerabilityDBPage() {
-  const [currentPage, setCurrentPage] = useState(1);
-  const itemsPerPage = 5; // 페이지 당 카드 수
-  const totalPages = Math.ceil(vulnerabilityDBDatas.length / itemsPerPage); // 카드 5개마다 1페이지
-
-  const pagesPerGroup = 10; //페이지 그룹 단위
-  const currentGroup = Math.ceil(currentPage / pagesPerGroup);
-  const startPage = (currentGroup - 1) * pagesPerGroup + 1;
-  const endPage = Math.min(startPage + pagesPerGroup - 1, totalPages);
-
-  // 현재 페이지에 해당하는 데이터만 가져오기
-  const currentData = vulnerabilityDBDatas.slice(
-    (currentPage - 1) * itemsPerPage,
-    currentPage * itemsPerPage,
-  );
-
   return (
     <div className="mx-auto mb-[1.188rem] mt-[1.688rem] flex w-[82.063rem] flex-col gap-[4.75rem]">
       <CardContainer />
-      <div className="flex gap-[6.375rem]">
-        {/* 취약점 DB */}
-        <VulnerabilityDB currentData={currentData} />
-        {/* 실시간 Topic */}
-        <RealTimeTopic />
-      </div>
-      <div className="flex justify-center">
-        <Pagination
-          currentPage={currentPage}
-          totalPages={totalPages}
-          startPage={startPage}
-          endPage={endPage}
-          setCurrentPage={setCurrentPage}
-        />
-      </div>
+      <Dashboard /> {/* 취약점 DB & 실시간 Topic */}
     </div>
   );
 }
@@ -228,78 +102,5 @@ function CardContainer() {
         );
       })}
     </div>
-  );
-}
-
-function VulnerabilityDB({
-  currentData,
-}: {
-  currentData: typeof vulnerabilityDBDatas;
-}) {
-  return (
-    <section>
-      <h2 className="text-2xl font-semibold leading-[1.816rem] tracking-[-0.01em]">
-        취약점 DB
-      </h2>
-      <div className="my-4">
-        <Label className="mr-3">HOT</Label>
-        <Label variant="unselected">NEW</Label>
-      </div>
-      <ul className="flex flex-col gap-4">
-        {currentData.map((data) => {
-          return (
-            <li key={data.id}>
-              <Card variant="article" size="long">
-                <CardHeader className="gap-0">
-                  <div className="mb-2 flex items-center gap-2">
-                    <Label>HOT</Label>
-                    <CardTitle
-                      size="small"
-                      color="black"
-                      className="line-clamp-1 font-normal"
-                    >
-                      {data.title}
-                    </CardTitle>
-                  </div>
-                  <CardSubTitle
-                    isSingleLine
-                    className="block truncate text-line-dark"
-                  >
-                    Microsoft
-                  </CardSubTitle>
-                </CardHeader>
-                <CardContent className="bg-purple-light">
-                  <span className="block truncate text-[#797979]">
-                    {data.content}
-                  </span>
-                </CardContent>
-                <CardFooter>
-                  <div className="flex gap-3">
-                    <IconPin width={28} height={28} />
-                    <IconExternalLink />
-                  </div>
-                  {/* createdAt의 날짜에서 현재 날짜 기준으로 계산 예정. 기능 추가할 때 수정하겠습니다! */}
-                  <CardSubTitle color="#A2A2A2">2일 전</CardSubTitle>
-                </CardFooter>
-              </Card>
-            </li>
-          );
-        })}
-      </ul>
-    </section>
-  );
-}
-
-function RealTimeTopic() {
-  return (
-    <section className="flex flex-col gap-4">
-      <h2 className="text-2xl font-semibold leading-[1.816rem] tracking-[-0.01em]">
-        실시간 Topic
-      </h2>
-      <p className="text-lg font-medium leading-[1.361rem] tracking-[-0.01em] text-gray-default">
-        03.08 10:00시 기준
-      </p>
-      <Ranking />
-    </section>
   );
 }
